Refactoring and verify token
